refactor: Rename `dockerfile_builders` package to `dockerfile`
