+++ internal/builders/dockerfile/bun_dockerfile.go
@@ -1,4 +1,4 @@
-package dockerfilebuilders
+package dockerfile
 
 import (
 	"fmt"
