+++ internal/builders/dockerfile/dockerfile_builder.go
@@ -1,4 +1,4 @@
-package dockerfilebuilders
+package dockerfile
 
 import (
 	"embed"
