+++ internal/builders/dockerfile/dotnet_dockerfile.go
@@ -1,4 +1,4 @@
-package dockerfilebuilders
+package dockerfile
 
 import (
 	"fmt"
@@ -10,7 +10,7 @@ import (
 	"github.com/jedib0t/go-pretty/v6/text"
 )
 
-func BuilderJavaGradleDockerfile(input DockerfileData) error {
+func BuildDotNetDockerfile(input DockerfileData) error {
 	if len(input.ProjectName) == 0 {
 		pN, err := setProjectName()
 		if err != nil {
@@ -23,31 +23,34 @@ func BuilderJavaGradleDockerfile(input DockerfileData) error {
 		fmt.Println(text.FgYellow.Sprintf("> This language doens't needs to specify the Entry File"))
 		os.Exit(1)
 	}
-	cmd := exec.Command("java", "-version")
-	javaVersionOutput, err := cmd.CombinedOutput()
+
+	cmd := exec.Command("dotnet", "--version")
+	dotnetVersionOutput, err := cmd.CombinedOutput()
 	if err != nil {
 		return err
 	}
-
-	versionPattern := `"(\d+)\.`
+	versionPattern := `(\d+)\.`
 	re := regexp.MustCompile(versionPattern)
 
-	matches := re.FindStringSubmatch(string(javaVersionOutput))
+	matches := re.FindStringSubmatch(string(dotnetVersionOutput))
 
 	if len(matches) < 2 {
-		return fmt.Errorf("failed to extract Java version number")
+		return fmt.Errorf("failed to extract Dotnet version number")
 	}
 	input.Version = matches[1]
 
-	datafile, err := templatesContent.ReadFile("templates/java_gradle_dockerfile.tmpl")
+	datafile, err := templatesContent.ReadFile("templates/dotnet_dockerfile.tmpl")
 	if err != nil {
 		return err
 	}
+
 	file, err := os.Create("Dockerfile")
 	if err != nil {
 		return err
 	}
+
 	defer file.Close()
+
 	builders.ApplyTemplate(file, string(datafile), input)
 
 	return nil
