+++ internal/builders/dockerfile/go_dockerfile.go
@@ -1,6 +1,7 @@
-package dockerfilebuilders
+package dockerfile
 
 import (
+	_ "embed"
 	"fmt"
 	"os"
 	"os/exec"
@@ -10,45 +11,45 @@ import (
 	"github.com/jedib0t/go-pretty/v6/text"
 )
 
-func BuildDotNetDockerfile(input DockerfileData) error {
+func BuildGoDockerfile(input DockerfileData) error {
 	if len(input.ProjectName) == 0 {
 		pN, err := setProjectName()
 		if err != nil {
 			os.Exit(1)
 		}
-
 		input.ProjectName = pN
 	}
+
 	if len(input.EntryFile) >= 1 {
 		fmt.Println(text.FgYellow.Sprintf("> This language doens't needs to specify the Entry File"))
 		os.Exit(1)
 	}
 
-	cmd := exec.Command("dotnet", "--version")
-	dotnetVersionOutput, err := cmd.CombinedOutput()
+	cmd := exec.Command("go", "version")
+	goVersionOutput, err := cmd.Output()
 	if err != nil {
 		return err
 	}
-	versionPattern := `(\d+)\.`
-	re := regexp.MustCompile(versionPattern)
 
-	matches := re.FindStringSubmatch(string(dotnetVersionOutput))
+	versionPattern := `go version go(\d+\.\d+\.\d+)`
+	versionRegex := regexp.MustCompile(versionPattern)
 
+	matches := versionRegex.FindStringSubmatch(string(goVersionOutput))
 	if len(matches) < 2 {
-		return fmt.Errorf("failed to extract Dotnet version number")
+		return fmt.Errorf("failed to extract Go version number")
 	}
-	input.Version = matches[1]
-
-	datafile, err := templatesContent.ReadFile("templates/dotnet_dockerfile.tmpl")
+	goVersion := matches[1]
+	datafile, err := templatesContent.ReadFile("templates/go_dockerfile.tmpl")
 	if err != nil {
 		return err
 	}
+
+	input.Version = goVersion
 
 	file, err := os.Create("Dockerfile")
 	if err != nil {
-		return err
+		return fmt.Errorf("Error creating Dockerfile: %s", err.Error())
 	}
-
 	defer file.Close()
 
 	builders.ApplyTemplate(file, string(datafile), input)
