+++ internal/builders/dockerfile/javascript_node_dockerfile.go
@@ -1,11 +1,9 @@
-package dockerfilebuilders
+package dockerfile
 
 import (
-	_ "embed"
 	"fmt"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"regexp"
 	"strings"
 
@@ -13,22 +11,20 @@ import (
 	"github.com/jedib0t/go-pretty/v6/text"
 )
 
-func BuildTypescriptNodeDockefile(input DockerfileData) error {
+func BuildJavascriptDockerfile(input DockerfileData) error {
 	if len(input.EntryFile) == 0 {
 		eF, err := setEntryfile()
 		if err != nil {
 			os.Exit(1)
 		}
 		input.EntryFile = eF
-
 	}
 
-	if strings.Contains(input.EntryFile, ".js") {
-		fmt.Println(text.FgYellow.Sprint("> Only allowed .ts files"))
+	if strings.Contains(input.EntryFile, ".ts") {
+		fmt.Println(text.FgYellow.Sprint("> Only allowed .js files"))
 
 		os.Exit(1)
 	}
-
 	if len(input.ProjectName) >= 1 {
 		fmt.Println(text.FgYellow.Sprintf("This language doens't needs to specify the Project Name"))
 		os.Exit(1)
@@ -50,18 +46,17 @@ func BuildTypescriptNodeDockefile(input DockerfileData) error {
 
 	nodeVersion := matches[1]
 
-	datafile, err := templatesContent.ReadFile("templates/typescript_node_dockerfile.tmpl")
+	datafile, err := templatesContent.ReadFile("templates/javascript_node_dockerfile.tmpl")
 	if err != nil {
 		return err
 	}
-
-	input.EntryFile = strings.TrimSuffix(input.EntryFile, filepath.Ext(input.EntryFile))
-	input.Version = nodeVersion
 
 	file, err := os.Create("Dockerfile")
 	if err != nil {
 		return fmt.Errorf("failed to creating Dockerfile: %s", err.Error())
 	}
+
+	input.Version = nodeVersion
 
 	defer file.Close()
