+++ internal/builders/dockerfile/rust_dockerfile.go
@@ -1,4 +1,4 @@
-package dockerfilebuilders
+package dockerfile
 
 import (
 	"fmt"
