+++ internal/builders/dockerfile/templates/dotnet_dockerfile.tmpl
