+++ internal/builders/dockerfile/templates/go_dockerfile.tmpl
