+++ internal/builders/dockerfile/templates/java_gradle_dockerfile.tmpl
