+++ internal/builders/dockerfile/templates/javascript_node_dockerfile.tmpl
