+++ internal/builders/dockerfile/templates/rust_dockerfile.tmpl
