+++ internal/builders/dockerfile/templates/tsx_bun_dockerfile.tmpl
