+++ internal/builders/dockerfile/templates/typescript_node_dockerfile.tmpl
